requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-appconf==1.0.5
django-compressor==2.2
django-cors-headers==3.11.0
django-libsass==0.7
djangorestframework==3.13.1
idna==3.3
libsass==0.21.0
Pillow==9.1.0
psycopg2==2.9.3
pytz==2022.1
rcssmin==1.0.6
requests==2.27.1
rjsmin==1.0.12
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
